requirements.txt
APScheduler==3.6.3
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
auth==0.5.3
beautifulsoup4==4.11.1
blinker==1.4
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-oauth-toolkit==2.0.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
drf-yasg==1.20.0
eventlet==0.33.0
falcon==3.1.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
gTTS==2.2.4
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.2
loguru==0.5.3
MarkupSafe==2.1.1
mongoengine==0.24.1
notify-py==0.3.3
notify2==0.3.1
numpy==1.22.3
oauthlib==3.2.0
openapi-codec==1.3.2
outcome==1.1.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
playsound==1.3.0
plyer==2.0.0
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==13.11
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.5
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.1
trio==0.20.0
trio-websocket==0.9.2
twilio==7.9.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
webdriver-manager==3.5.4
Werkzeug==2.1.2
whitenoise==6.1.0
win32-setctime==1.1.0
wrapt==1.14.1
wsproto==1.1.0
X9GCBSpgF7k5ujmWOxBt1CNDjOpQ7ldTRTpaNIigEZ1ja5jrEdb4kFQZIm6mGVxfNz8kItddlpyaV39wGpE4i4ppaYue7jbd6kF5bB5Q3dfUfRGSSiSuGp8fMRGq3Nfs


curl -X POST -d "client_id=DoPyPM7A6b36ruDLUYjjbNVDGnpsbYrBKoZ1aGIX&client_secret=X9GCBSpgF7k5ujmWOxBt1CNDjOpQ7ldTRTpaNIigEZ1ja5jrEdb4kFQZIm6mGVxfNz8kItddlpyaV39wGpE4i4ppaYue7jbd6kF5bB5Q3dfUfRGSSiSuGp8fMRGq3Nfs&grant_type=password&username=roman&password=Dadefo95" http://127.0.0.1:8000/auth/token